requirements.txt
streamlit
openai
numpy
torch
requests
urllib3
python-dotenv  # Optional, if using dotenv to manage environment variables
